regcoach_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv>=1.0.1
numpy>=1.26.4,<2
google-generativeai==0.8.5
pydantic>=2.7,<3
httpx>=0.27.0
python-multipart>=0.0.6
